refactor: TicketSize enum 클래스, utils 패키지로 이동
