+++ src/main/java/lotto/domain/ticket/validation/TicketValidator.java
@@ -3,9 +3,9 @@ package lotto.domain.ticket.validation;
 import java.util.Collections;
 import java.util.List;
 
-import lotto.domain.ticket.condition.TicketSize;
 import lotto.exception.LottoException;
 import lotto.exception.LottoExceptionStatus;
+import lotto.utils.TicketSize;
 
 public class TicketValidator {
